requirements.txt
# MQTT-Redis Data Ingestion System Requirements

# Core dependencies
paho-mqtt>=1.6.1,<2.0.0
redis>=4.5.0,<6.0.0

# JSON handling and configuration
# (json is part of Python standard library)

# Logging and utilities
# (logging is part of Python standard library)

# Type hints
typing-extensions>=4.0.0; python_version < "3.8"

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Testing MQTT broker (for development/testing)
# mosquitto can be installed separately

# Optional: Enhanced Redis support
hiredis>=2.0.0

# Optional: Advanced logging
colorlog>=6.7.0

# Documentation
# sphinx>=5.0.0
# sphinx-rtd-theme>=1.2.0